Moves "labels" to inside "fieldsConfig"
We shouldn't hard code that there will be 2 labels, since this component can receive any number of fields. It makes more sense to have each field configuration contain the label to be used, if there is one.

Also, it's best to render the "label" element even if no label is given, so that labels can be correctly placed even if only given for the last fields.